requirements.txt
beautifulsoup4==4.13.5
chronos==0.3
matplotlib==3.8.2
mistralai==1.9.10
networkx==3.5
numpy==2.3.3
pandas==2.3.2
Requests==2.32.5
seaborn==0.13.2
streamlit==1.47.0
tavily_python==0.5.1
torch==2.6.0
